complete merge and fix a bug in newton/bisection choice logic
