+++ src/serac/numerics/functional/tensor.hpp
@@ -1431,140 +1431,6 @@ SERAC_HOST_DEVICE constexpr auto chop(const tensor<double, m, n>& A)
   return copy;
 }
 
-
-template <typename T>
-struct SolverResults {
-  T root;
-  bool converged;
-  unsigned int iterations;
-
-  SolverResults(T root, bool converged, unsigned int iterations) :
-    root(root), converged(converged), iterations(iterations) {};
-};
-
-template <typename function, typename... ParamTypes>
-auto solve_scalar_equation(function && f, double x0, double tolerance,
-                             double lower_bound, double upper_bound, ParamTypes... params)
-{
-  double x, df_dx;
-  double fl = f(lower_bound, get_value(params)...);
-  double fh = f(upper_bound, get_value(params)...);
-
-  if (fl*fh > 0) {
-    SLIC_ERROR("solve_scalar_equation: root not bracketed by input bounds.");
-  }
-  
-
-  unsigned int iterations = 0;
-  bool converged = false;
-
-  // handle corner case where one of the brackets is the root
-  if (fl == 0) {
-    x = lower_bound;
-    converged = true;
-  } else if (fh == 0) {
-    x = upper_bound;
-    converged = true;
-  }
-  
-  if (converged) {
-    
-    df_dx = get_gradient(f(make_dual(x), get_value(params)...));
-
-  } else {
-
-    // orient search so that f(xl) < 0
-    double xl = lower_bound;
-    double xh = upper_bound;
-    if (fl > 0) {
-      xl = upper_bound;
-      xh = lower_bound;
-    }
-
-    // move initial guess if it is not between brackets
-    if (x0 < lower_bound || x0 > upper_bound) {
-      x0 = 0.5*(lower_bound + upper_bound);
-    }
-    const unsigned int MAX_ITERATIONS = 25;
-    x = x0;
-    double dx_old = std::abs(upper_bound - lower_bound);
-    double dx = dx_old;
-    auto R = f(make_dual(x), get_value(params)...);
-    auto fval = get_value(R);
-    df_dx = get_gradient(R);
-
-    while (!converged) {
-      if (iterations == MAX_ITERATIONS) {
-        SLIC_WARNING("solve_scalar_equation failed to converge in allotted iterations.");
-        break;
-      }
-      
-      // use bisection if Newton oversteps brackets or is not decreasing sufficiently
-      if ((x - xh)*df_dx - fval > 0 ||
-          (x - xl)*df_dx - fval < 0 ||
-          std::abs(2.*fval) > dx_old*df_dx) {
-        dx_old = dx;
-        dx = 0.5*(xh - xl);
-        x = xl + dx;
-        converged = (x == xl);
-      } else { // use Newton step
-        dx_old = dx;
-        dx = fval/df_dx;
-        auto temp = x;
-        x -= dx;
-        converged = (x == temp);
-      }
-
-      // std::cout << "iter " << i << " x = " << x << std::endl;
-
-      // convergence check
-      converged = converged || (std::abs(dx) < tolerance);
-      
-      // function and jacobian evaluation
-      R = f(make_dual(x), get_value(params)...);
-      fval = get_value(R);
-      df_dx = get_gradient(R);
-
-      // maintain bracket on root
-      if (fval < 0) {
-        xl = x;
-      } else {
-        xh = x;
-      }
-
-      ++iterations;
-    }
-
-  }
-
-  // Accumulate derivatives so that the user can get derivatives
-  // with respect to parameters, subject to constraing that f(x, p) = 0 for all p
-  // Conceptually, we're doing the following:
-  // [fval, df_dp] = f(get_value(x), p)
-  // df = 0
-  // for p in params:
-  //   accumulate df += inner(df_dp, p_dot)
-  // x_dot = -df / df_dx
-  constexpr bool contains_duals = (is_dual_number<ParamTypes>::value || ...);
-  if constexpr (contains_duals) {
-    auto seq = std::make_integer_sequence<int, static_cast<int>(sizeof...(params))>();
-    auto df = solver_derivative_helper(f, x, tuple{params...}, seq);
-    return SolverResults(dual<double>{x, -df/df_dx}, converged, iterations);
-  }
-  if constexpr (!contains_duals) {
-
-    return SolverResults(x, converged, iterations);
-  }
-}
-
-template <typename function, typename... ParamTypes, int... i>
-auto solver_derivative_helper(function && f, double x, tuple<ParamTypes...> params, std::integer_sequence<int, i...>)
-{
-  auto [fval, df_dp] = f(get_value(x), get<i>(params)...);
-  return (inner(get<i>(df_dp), get_gradient(get<i>(params))) + ...);
-}
-
-
 /// @cond
 namespace detail {
