+++ src/serac/numerics/functional/tests/test_newton.cpp
@@ -6,7 +6,7 @@
 
 #include <gtest/gtest.h>
 
-#include "serac/numerics/functional/tensor.hpp"
+#include "serac/numerics/functional/tuple_tensor_dual_functions.hpp"
 
 using namespace serac;
 
@@ -68,61 +68,6 @@ TEST(ScalarEquationSolver, DerivativeOfPrimal)
     EXPECT_NEAR(dsqrt_p, 0.5/std::sqrt(p), 1e-12);
 }
 
-TEST(ScalarEquationSolver, DerivativeWithTensorParameter)
-{
-    auto my_norm = [](auto A) {
-        auto p = squared_norm(A);
-        double tolerance = 1e-10;
-        double lower = 1e-3;
-        double upper = 20.0;
-        double x0 = get_value(p);
-        return solve_scalar_equation([](auto x, auto a){ return g(x, a); }, x0, tolerance, lower, upper, p);
-    };
-
-    tensor< double, 3, 3 > A = {{{1.0, 2.0, 3.0}, {4.0, 5.0, 6.0}, {7.0, 8.0, 9.0}}};
-
-    auto [my_norm_p, dmy_norm_p] = my_norm(make_dual(A));
-    auto [norm_p, dnorm_p] = norm(make_dual(A));
-    std::cout << "my_norm_p  = " << my_norm_p << std::endl;
-    std::cout << "norm_p     = " << norm_p << std::endl;
-    std::cout << "dmy_norm_p = " << dmy_norm_p << std::endl;
-    std::cout << "dnorm_p    = " << dnorm_p << std::endl;
-}
-
-TEST(ScalarEquationSolver, DerivativeWithTensorParameter2)
-{
-    auto my_norm = [](auto A) {
-        double tolerance = 1e-10;
-        double lower = 1e-3;
-        double upper = 20.0;
-        double x0 = 10.0;
-        return solve_scalar_equation([](auto x, auto P){ return x*x - squared_norm(P); }, x0, tolerance, lower, upper, A);
-    };
-
-    tensor< double, 3, 3 > A = {{{1.0, 2.0, 3.0}, {4.0, 5.0, 6.0}, {7.0, 8.0, 9.0}}};
-
-    auto [my_norm_p, dmy_norm_p] = my_norm(make_dual(A));
-    auto [norm_p, dnorm_p] = norm(make_dual(A));
-    std::cout << "my_norm_p  = " << my_norm_p << std::endl;
-    std::cout << "norm_p     = " << norm_p << std::endl;
-    std::cout << "dmy_norm_p = " << dmy_norm_p << std::endl;
-    std::cout << "dnorm_p    = " << dnorm_p << std::endl;
-}
-
-    // def test_scalar_newton_jvp(self):
-    //     def my_sqrt(p):
-    //         tol = 1e-12
-    //         x = newton.solve_scalar(f, p, tol, p)
-    //         return x
-        
-    //     p = 2.0
-    //     s = my_sqrt(p)
-    //     print(f"{s:.14f}")
-    //     dsqrt = jax.jacfwd(my_sqrt)
-    //     print(f"{dsqrt(p):.14f}")
-    //     error = np.abs(dsqrt(p) - 0.5/np.sqrt(p))
-    //     self.assertLess(error, 1e-10)
-
 TEST(ScalarEquationSolver, AbortsIfRootNotBracketedByCaller)
 {
     double x0 = 5.0;
@@ -183,19 +128,49 @@ TEST(ScalarEquationSolver, ConvergesWithGuessOutsideNewtonBasin)
     EXPECT_LT(error, tolerance);
 }
 
-TEST(ScalarEquationSolver, DerivativeOfPrimal)
+// TEST(ScalarEquationSolver, DerivativeWithTensorParameter)
+// {
+//     auto my_norm = [](auto A) {
+//         auto p = squared_norm(A);
+//         double tolerance = 1e-10;
+//         double lower = 1e-3;
+//         double upper = 20.0;
+//         double x0 = get_value(p);
+//         return solve_scalar_equation([](auto x, auto a){ return g(x, a); }, x0, tolerance, lower, upper, p);
+//     };
+
+//     tensor< double, 3, 3 > A = {{{1.0, 2.0, 3.0}, {4.0, 5.0, 6.0}, {7.0, 8.0, 9.0}}};
+
+//     auto [my_norm_p, dmy_norm_p] = my_norm(make_dual(A));
+//     auto [norm_p, dnorm_p] = norm(make_dual(A));
+//     std::cout << "my_norm_p  = " << my_norm_p << std::endl;
+//     std::cout << "norm_p     = " << norm_p << std::endl;
+//     std::cout << "dmy_norm_p = " << dmy_norm_p << std::endl;
+//     std::cout << "dnorm_p    = " << dnorm_p << std::endl;
+// }
+
+TEST(ScalarEquationSolver, DerivativeWithTensorParameter2)
 {
-    auto my_sqrt = [](dual<double> p) {
-        double x0 = get_value(p);
-        double tolerance = 1e-6;
-        double lower = 0;
-        double upper = (get_value(p) > 1.0) ? get_value(p) : 1.0;
-        auto result = solve_scalar_equation([](auto x, auto a){ return g(x, a); }, x0, tolerance, lower, upper, p);
-        return result.root;
+    auto my_norm = [](auto A) {
+        double tolerance = 1e-10;
+        double lower = 1e-3;
+        double upper = 20.0;
+        double x0 = 10.0;
+        auto sol = solve_scalar_equation([](auto x, auto P){ return x*x - squared_norm(P); }, x0, tolerance, lower, upper, A);
+        return sol.root;
     };
-    double p = 2.0;
-    auto [sqrt_p, dsqrt_p] = my_sqrt(make_dual(p));
-    EXPECT_NEAR(dsqrt_p, 0.5/std::sqrt(p), 1e-12);
+
+    tensor< double, 3, 3 > A = {{{1.0, 2.0, 3.0}, 
+                                 {4.0, 5.0, 6.0}, 
+                                 {7.0, 8.0, 9.0}}};
+
+    // auto [my_norm_p, dmy_norm_p] = my_norm(make_dual(A));
+    auto my_norm_p = my_norm(make_dual(A));
+    auto [norm_p, dnorm_p] = norm(make_dual(A));
+    std::cout << "my_norm_p  = " << my_norm_p << std::endl;
+    std::cout << "norm_p     = " << norm_p << std::endl;
+    //std::cout << "dmy_norm_p = " << dmy_norm_p << std::endl;
+    std::cout << "dnorm_p    = " << dnorm_p << std::endl;
 }
 
 int main(int argc, char* argv[])
